Adding support for ST7789 width and height
Modified existing code to allow for the width and height to optionally be specified for ST7789 displays, as it was hardcoded to 240x240. Tested on a 240x320 and 240x240 ST7789 display.

`def display` in the st7789 class still references 240 for `w` and `h`, but despite this it doesn't appear to affect the display on a larger 240x320 display.